# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Project-HarvestNet Open Data Exchange API
  description: |
    BRICS AgriN Digital Public Infrastructure — open, sovereign Agri Data Exchange.

    This API serves as the Layer 4 (L4) public interface for Project-HarvestNet,
    exposing regenerative crop advisory, disease diagnosis, regional outbreak alerts,
    farmer identity management, land plot mapping, and DIY soil testing.

    **Novelty differentiators:**
    - N1: Deterministic rules engine — the LLM never decides agronomy
    - N2: Regional outbreak early-warning signal from aggregated diagnosis events
    - N3: Confidence-gated human escalation for uncertain diagnoses
  version: 1.0.0
  contact:
    name: Project-HarvestNet Team
servers:
  - url: http://localhost:3000
    description: Local development

# ═══════════════════════════════════════════════════════════════
# Section 8 (v2) + Addendum Section 23 (v3) paths
# ═══════════════════════════════════════════════════════════════
paths:
...
  /api/v1/farmers/register:
    post:
      summary: Register a farmer and issue a farmer_id
      description: Creates a new FarmerProfile with consent (Addendum Section 18).
      tags: [Farmer Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, phone_number, preferred_language, consent]
              properties:
                display_name:
                  type: string
                  example: "Raju Patil"
                phone_number:
                  type: string
                  description: Raw phone number — stored as hash by identity-api
                preferred_language:
                  type: string
                  example: "kn-IN"
                home_region_geohash5km:
                  type: string
                consent:
                  $ref: "#/components/schemas/FarmerConsentInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  farmer_id:
                    type: string
                  registered_at:
                    type: string
                    format: date-time
  /api/v1/farmers/{farmer_id}:
    get:
      summary: Fetch a farmer profile
      tags: [Farmer Identity]
      parameters:
        - name: farmer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: FarmerProfile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerProfile"
        "404":
          description: Farmer not found
    patch:
      summary: Update profile or consent
      tags: [Farmer Identity]
      parameters:
        - name: farmer_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                preferred_language:
                  type: string
                consent:
                  $ref: "#/components/schemas/FarmerConsentInput"
      responses:
        "200":
          description: Updated FarmerProfile
  /api/v1/farmers/{farmer_id}/plots:
    post:
      summary: Save a drawn land-plot boundary
      description: |
        Accepts a GeoJSON Polygon boundary. Returns plot_id with derived
        area_hectares (Shoelace formula) and centroid (arithmetic mean).
      tags: [Land Plots]
      parameters:
        - name: farmer_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boundary]
              properties:
                boundary:
                  $ref: "#/components/schemas/GeoJSONPolygon"
                label:
                  type: string
                  example: "north field"
      responses:
        "201":
          description: Plot created
          content:
            application/json:
              schema:
                type: object
                properties:
                  plot_id:
                    type: string
                  area_hectares:
                    type: number
                  centroid:
                    type: object
                    properties:
                      latitude:
                        type: number
                      longitude:
                        type: number
    get:
      summary: List a farmer's plots
      tags: [Land Plots]
      parameters:
        - name: farmer_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of LandPlot objects
  /api/v1/plots/{plot_id}:
    get:
      summary: Fetch a plot plus latest NDVI/soil snapshot
      description: |
        Returns plot data with a spatial snapshot proxied from
        data-aggregator using the plot's centroid coordinates.
      tags: [Land Plots]
      parameters:
        - name: plot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: LandPlot with spatial snapshot
        "404":
          description: Plot not found
...
components:
  schemas:
...
    FarmerProfile:
      type: object
      description: Farmer identity profile (Addendum 18.3)
      properties:
        farmer_id:
          type: string
        registered_at:
          type: string
          format: date-time
        display_name:
          type: string
        phone_number_hash:
          type: string
        preferred_language:
          type: string
        home_region_geohash5km:
          type: string
        consent:
          $ref: "#/components/schemas/FarmerConsent"
        plot_ids:
          type: array
          items:
            type: string
    FarmerConsent:
      type: object
      properties:
        consent_version:
          type: string
        data_sharing_with_partners:
          type: boolean
        climate_trend_aggregation:
          type: boolean
        recorded_at:
          type: string
          format: date-time
    FarmerConsentInput:
      type: object
      required: [data_sharing_with_partners]
      properties:
        consent_version:
          type: string
          default: "1.0"
        data_sharing_with_partners:
          type: boolean
        climate_trend_aggregation:
          type: boolean
          default: false
    GeoJSONPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
...